rename: 型定義ファイルの命名を変更
