+++ src/components/Title/index.tsx
@@ -1,5 +1,4 @@
 import { Size } from '@/types/button';
-import styles from './styles.module.scss';
 import { CSSProperties } from 'react';
 import { FONT_SIZE_MAP } from './const';
 
@@ -13,11 +12,7 @@ function Title({ text, size = 'md' }: Props) {
     ...FONT_SIZE_MAP[size],
   };
 
-  return (
-    <h2 style={titleStyle} className={styles.title}>
-      {text}
-    </h2>
-  );
+  return <h2 style={titleStyle}>{text}</h2>;
 }
 
 export default Title;
