+++ src/types/button.d.ts
@@ -1,0 +1,2 @@
+export type Size = 'xs' | 'sm' | 'md' | 'lg' | 'xl';
+export type ButtonType = 'normal' | 'dark' | 'outline';
