--- src/types/size.d.ts
@@ -1,1 +1,0 @@
-export type Size = 'xs' | 'sm' | 'md' | 'lg' | 'xl';
